requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0

# Trading & Market Data
ccxt==4.2.25
pandas==2.1.4
numpy==1.26.3
ta==0.11.0
yfinance==0.2.35

# Machine Learning & AI
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
xgboost==2.0.3

# Technical Analysis
pandas-ta==0.3.14b0
mplfinance==0.12.10b0

# Data Storage
pymongo==4.6.1
redis==5.0.1

# Voice Processing
speechrecognition==3.10.1
pydub==0.25.1
gtts==2.4.0

# Utilities
requests==2.31.0
aiohttp==3.9.1
schedule==1.2.0
pytz==2023.3
colorama==0.4.6

# Visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0

# Testing & Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0